+++ DawnLib.Dusk/src/API/Definitions/Dungeon/DungeonData.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace Dusk;
-
-[Serializable]
-public class DungeonData : EntityData<DuskAdditionalTilesReference>
-{
-}